refactor: delete the etcd detach from the greptimedb

+++ cmd/app/cluster/delete/delete.go
@@ -12,8 +12,9 @@ import (
 )
 
 type deleteClusterCliOptions struct {
-	Namespace    string
-	TearDownEtcd bool
+	Namespace     string
+	TearDownEtcd  bool
+	ETCDNamespace string
 }
 
 func NewDeleteClusterCommand(l log.Logger) *cobra.Command {
@@ -29,6 +30,7 @@ func NewDeleteClusterCommand(l log.Logger) *cobra.Command {
 			}
 
 			clusterName, namespace := args[0], options.Namespace
+			clusterExist := true
 			l.Infof("⚠️ Deleting cluster '%s' in namespace '%s'...\n", log.Bold(clusterName), log.Bold(namespace))
 
 			m, err := manager.New(l, false)
@@ -41,23 +43,34 @@ func NewDeleteClusterCommand(l log.Logger) *cobra.Command {
 				ClusterName: clusterName,
 				Namespace:   options.Namespace,
 			})
-			if err != nil && errors.IsNotFound(err) {
+			if errors.IsNotFound(err) {
 				l.Infof("Cluster '%s' in '%s' not found\n", clusterName, namespace)
-				return nil
-			} else if err != nil {
+				clusterExist = false
+			} else if err != nil && !errors.IsNotFound(err) {
 				return err
 			}
 
-			if err := m.DeleteCluster(ctx, &manager.DeleteClusterOption{
-				ClusterName:  clusterName,
-				Namespace:    options.Namespace,
-				TearDownEtcd: options.TearDownEtcd,
-			}); err != nil {
-				return err
+			if clusterExist {
+				if err := m.DeleteCluster(ctx, &manager.DeleteClusterOption{
+					ClusterName: clusterName,
+					Namespace:   options.Namespace,
+				}); err != nil {
+					return err
+				}
+
+				// TODO(zyy17): Should we wait until the cluster is actually deleted?
+				l.Infof("Cluster '%s' in namespace '%s' is deleted!\n", clusterName, namespace)
 			}
 
-			// TODO(zyy17): Should we wait until the cluster is actually deleted?
-			l.Infof("Cluster '%s' in namespace '%s' is deleted!\n", clusterName, namespace)
+			if options.TearDownEtcd {
+				l.Infof("⚠️ Deleting etcd cluster in namespace '%s'...\n", log.Bold(options.ETCDNamespace))
+				if err := m.DeleteETCDCluster(ctx, &manager.DeleteETCDClusterOption{
+					Namespace: options.ETCDNamespace,
+				}); err != nil {
+					return err
+				}
+				l.Infof("ETCD cluster in namespace '%s' is deleted!\n", options.ETCDNamespace)
+			}
 
 			return nil
 		},
@@ -65,6 +78,7 @@ func NewDeleteClusterCommand(l log.Logger) *cobra.Command {
 
 	cmd.Flags().StringVarP(&options.Namespace, "namespace", "n", "default", "Namespace of GreptimeDB cluster.")
 	cmd.Flags().BoolVar(&options.TearDownEtcd, "tear-down-etcd", false, "Tear down etcd cluster.")
+	cmd.Flags().StringVar(&options.ETCDNamespace, "etcd-namespace", "default", "Namespace of etcd cluster.")
 
 	return cmd
 }
